requirements.txt
# Data Engineering
pandas==2.2.0
numpy==1.26.4
kaggle==1.6.6
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0          # AWS S3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # Supabase/PostgreSQL
pymysql==1.1.0          # MySQL fallback

# Data Validation
pydantic==2.5.3

# ML Models
scikit-learn==1.4.0
xgboost==2.0.3
tensorflow==2.15.0
shap==0.44.0

# API
fastapi==0.109.0
uvicorn==0.27.0

# Dashboard
streamlit==1.31.0
plotly==5.18.0

# LLM Explainability
google-generativeai==0.3.2   # Gemini API (free tier)

# Testing
pytest==7.4.4
